feat: reorganized to prevent header rerender on app select
